refactor: move sessionCaptureWriter to internal/agent
Hoist the JSONL session-capture writer out of the daemon package so
both the daemon worker and the foreground roborev fix CLI can use it
without a daemon dependency. Behavior is unchanged; the daemon worker
keeps the same callback signature.
