+++ internal/agent/session_writer.go
@@ -1,0 +1,98 @@
+package agent
+
+import (
+	"bytes"
+	"io"
+	"sync"
+)
+
+// SessionCaptureWriter is a transparent passthrough writer that scans
+// the byte stream for the first agent session ID JSON event
+// (session_id / sessionId / sessionID / type:session / thread.started)
+// and surfaces it via SessionID(). After capture, scanning stops and
+// the writer continues to forward bytes unchanged.
+//
+// Use Flush() once writes are complete so any trailing partial line
+// is also examined.
+type SessionCaptureWriter struct {
+	dst       io.Writer
+	onCapture func(string)
+	mu        sync.Mutex
+	buf       bytes.Buffer
+	sessionID string
+}
+
+// NewSessionCaptureWriter wraps dst and reports the first captured
+// session ID via onCapture (if non-nil). Bytes pass through to dst
+// unchanged.
+func NewSessionCaptureWriter(dst io.Writer, onCapture func(string)) *SessionCaptureWriter {
+	return &SessionCaptureWriter{dst: dst, onCapture: onCapture}
+}
+
+func (w *SessionCaptureWriter) Write(p []byte) (int, error) {
+	n, err := w.dst.Write(p)
+	if n > 0 {
+		w.capture(p[:n])
+	}
+	return n, err
+}
+
+// Flush examines any buffered partial line. Call this before reading
+// SessionID() to ensure capture is complete.
+func (w *SessionCaptureWriter) Flush() {
+	var captured string
+
+	w.capture(nil)
+
+	w.mu.Lock()
+	if w.sessionID == "" && w.buf.Len() > 0 {
+		if sessionID := ExtractSessionID(w.buf.String()); sessionID != "" {
+			w.sessionID = sessionID
+			captured = sessionID
+		}
+	}
+	w.mu.Unlock()
+
+	if captured != "" && w.onCapture != nil {
+		w.onCapture(captured)
+	}
+}
+
+// SessionID returns the first captured session ID, or "" if none.
+func (w *SessionCaptureWriter) SessionID() string {
+	w.mu.Lock()
+	defer w.mu.Unlock()
+	return w.sessionID
+}
+
+func (w *SessionCaptureWriter) capture(p []byte) {
+	var captured string
+
+	w.mu.Lock()
+	if w.sessionID != "" {
+		w.mu.Unlock()
+		return
+	}
+	if len(p) > 0 {
+		_, _ = w.buf.Write(p)
+	}
+	for w.sessionID == "" {
+		data := w.buf.Bytes()
+		idx := bytes.IndexByte(data, '\n')
+		if idx < 0 {
+			break
+		}
+		line := string(data[:idx+1])
+		w.buf.Next(idx + 1)
+		if sessionID := ExtractSessionID(line); sessionID != "" {
+			w.sessionID = sessionID
+			w.buf.Reset()
+			captured = sessionID
+		}
+	}
+	w.mu.Unlock()
+
+	if captured != "" && w.onCapture != nil {
+		w.onCapture(captured)
+	}
+}
